nw Co-authored-by: noscur <noscur@users.noreply.github.com> Co-authored-by: Lysio4 <Lysio4@users.noreply.github.com>

+++ utils/manip.c
@@ -143,7 +143,7 @@ char *remove_dot_segments(const char *input) {
     return output;
 }
 bool isin(char* str, char* list[]){
-	for(int i=0; i<sizeof(list)/sizeof(list[0]); i++){
+	for(long unsigned int i=0; i<sizeof(list)/sizeof(list[0]); i++){
 		if(strcmp(str, list[i])==0){
 			return true;
 		}
